+++ src/tests/index.js
@@ -1,0 +1,14 @@
+import * as creationTests from "./suites/creationTests";
+import * as cleanupTests from "./suites/cleanupTests";
+import * as catchUpTests from "./suites/catchUpTests";
+import * as mergeTests from "./suites/mergeTests";
+import * as switchVersionTests from "./suites/switchVersionTests";
+
+
+export {
+    creationTests,
+    cleanupTests,
+    catchUpTests,
+    mergeTests,
+    switchVersionTests
+}
