+++ src/tests/suites/switchVersionTests.js
@@ -1,0 +1,30 @@
+import { strict as assert } from 'assert';
+import { runCreateSaga } from "../../saga/create";
+import { runOperation } from "../../saga/runOperation";
+import { item, TEST_URL } from "../../constants";
+import { getItemRangeValues } from "../testHelpers";
+import { getGlobal } from "../../commands/commands";
+
+
+export async function testSwitchVersions() {
+    
+    // First, we create the project
+    await runCreateSaga(TEST_URL, "email");
+
+    // Then, we make sure the personal branch is checked out
+    const head = (await runOperation(getItemRangeValues, item.HEAD))[0][0];
+    assert.equal("email", head, "Personal branch should be checked out");
+
+    // Then, we switch versions
+    const g = getGlobal();
+    await g.switchVersion();
+
+    const newHead = (await runOperation(getItemRangeValues, item.HEAD))[0][0];
+    assert.equal("master", newHead, "Master branch should be checked out");
+
+    await g.switchVersion();
+    const newNewHead = (await runOperation(getItemRangeValues, item.HEAD))[0][0];
+    assert.equal("email", newNewHead, "Personal branch should be checked out again");
+
+    return true;
+}
